Add files via upload
Lowestprice is a working file giving the place which is the lowest price. Is returns name, price and provider in json format from post. 

Proces does a variety of things, it can search for an item and return in array format. It can add an item to user basket. It can also return users basket. However GET and POST need to be corrected.